+++ src/main/java/com/alibaba/excel/converters/shortconverter/ShortBooleanConverter.java
@@ -17,7 +17,7 @@ public class ShortBooleanConverter implements Converter<Short> {
 
     @Override
     public Class supportJavaTypeKey() {
-        return Long.class;
+        return Short.class;
     }
 
     @Override
